Completed MVP definition of BallDetector class, now need to implement in .cpp file.
